Move reward config to intent schema
Prior to this, the reward config was set as a constant with option of
overriding by passing it as a parameter to the evaluator functions. Now,
we expect the reward config to be a part of the intent schema, which
needs to be provided during initialisation of the evaluator module.